Avoid an infinite loop on unknown Linux distributions
The code to handle Linux looks something like this:

while (1)
    if is Alpine Linux
        break
    else if is Arch Linux
        break

The problem is that if no distribution matches, the loop will not end.
This just removes the loop, while adding a default "unsupported" ASCII
art and package count. As a result, I've also removed the "else" block
that catches unsupported operating systems since it will use the same
defeault image that unknown Linux distributions now use.